refactor(compile): extract zksync methods
